+++ tiledesk-chatbot-util/test/test_td_button_tag.js
@@ -54,6 +54,53 @@ describe('TiledeskChatbotUtil', function() {
 
 describe('TiledeskChatbotUtil', function() {
     describe('parseReply() of tdButton', function() {
+        it('should return an intro text with a couple of buttons - 2', function() {
+            // const cbutil = new TiledeskChatbotUtil();
+            // const text = 'Intro text\ntdButton:Button 1\ntdButton:Button 2';
+            const text = 'Grazie Giulio, il tuo indirizzo email andrea@gmail.xi è corretto. A breve riceverai un messaggio con tutte le informazioni necessarie per utilizzare il tuo conto corrente. A presto! 🙂\n* Grazie\n* Vorrei parlare con un operatore';
+            console.log("parsing text:", text);
+            const reply = TiledeskChatbotUtil.parseReply(text);
+            console.log("reply buttons:", JSON.stringify(reply));
+            assert(reply.message != null);
+            assert(reply.message.text != null);
+            assert.strictEqual(reply.message.type, TiledeskChatbotUtil.TYPE_TEXT);
+            assert.strictEqual(reply.message.text, 'Grazie Giulio, il tuo indirizzo email andrea@gmail.xi è corretto. A breve riceverai un messaggio con tutte le informazioni necessarie per utilizzare il tuo conto corrente. A presto! 🙂');
+            assert(reply.message.attributes != null);
+            assert(reply.message.attributes.attachment != null);
+            assert(reply.message.attributes.attachment.type != null);
+            assert.strictEqual(reply.message.attributes.attachment.type, 'template');
+            assert(reply.message.attributes.attachment.buttons != null);
+            assert(reply.message.attributes.attachment.buttons.length == 2);
+            assert.strictEqual(reply.message.attributes.attachment.buttons[0].type, TiledeskChatbotUtil.TYPE_BUTTON_TEXT);
+            assert.strictEqual(reply.message.attributes.attachment.buttons[0].value, 'Grazie');
+            assert.strictEqual(reply.message.attributes.attachment.buttons[1].type, TiledeskChatbotUtil.TYPE_BUTTON_TEXT);
+            assert.strictEqual(reply.message.attributes.attachment.buttons[1].value, 'Vorrei parlare con un operatore');
+            
+            // MESSAGE:
+            // {
+            //     "message": {
+            //         "text": "Intro text",
+            //         "type": "text",
+            //         "attributes": {
+            //             "attachment": {
+            //                 "type": "template",
+            //                 "buttons": [{
+            //                     "type": "text",
+            //                     "value": "Button 1"
+            //                 }, {
+            //                     "type": "text",
+            //                     "value": "Button 2"
+            //                 }]
+            //             }
+            //         }
+            //     }
+            // }
+        });
+    });
+});
+
+describe('TiledeskChatbotUtil', function() {
+    describe('parseReply() of tdButton', function() {
         it('should return an intro text with four buttons with mixed spaces between the * and the button text', function() {
             // const cbutil = new TiledeskChatbotUtil();
             // const text = 'Intro text\ntdButton:Button 1\ntdButton:Button 2';
@@ -271,7 +318,7 @@ describe('TiledeskChatbotUtil', function() {
 // //             //                 "buttons": [{
 // //             //                     "type": "url",
 // //             //                     "value": "Button with text",
-// //             //                     "link": "http://www.google.com",
+ // //             //                     "link": "http://www.google.com",
 // //             //                     "target": "blank"
 // //             //                 }]
 // //             //             }
@@ -286,8 +333,8 @@ describe('TiledeskChatbotUtil', function() {
     describe('parseReply() of tdLinkParent buttons', function() {
         it('should return a link.PARENT button', function() {
             // const cbutil = new TiledeskChatbotUtil();
-            const text = 'Intro text\n* Button with text  http://www.google.com';
-            // const text = 'Intro text\n* Button with text > http://www.google.com';
+            //const text = 'Intro text\n* Button with text  http://www.google.com';
+            const text = 'Intro text\n* Button with text link in parent < http://www.google-in-parent.com';
             
             console.log("parsing text:", text);
             const reply = TiledeskChatbotUtil.parseReply(text);
@@ -301,8 +348,8 @@ describe('TiledeskChatbotUtil', function() {
             assert(reply.message.attributes.attachment.buttons != null);
             assert(reply.message.attributes.attachment.buttons.length == 1);
             assert.strictEqual(reply.message.attributes.attachment.buttons[0].type, TiledeskChatbotUtil.TYPE_BUTTON_URL);
-            assert.strictEqual(reply.message.attributes.attachment.buttons[0].value, 'Button with text');
-            assert.strictEqual(reply.message.attributes.attachment.buttons[0].link, 'http://www.google.com');
+            assert.strictEqual(reply.message.attributes.attachment.buttons[0].value, 'Button with text link in parent');
+            assert.strictEqual(reply.message.attributes.attachment.buttons[0].link, 'http://www.google-in-parent.com');
             assert.strictEqual(reply.message.attributes.attachment.buttons[0].target, TiledeskChatbotUtil.TARGET_BUTTON_LINK_PARENT);
             // MESSAGE:
             // {
@@ -317,6 +364,50 @@ describe('TiledeskChatbotUtil', function() {
             //                     "value": "Button with text",
             //                     "link": "http://www.google.com",
             //                     "target": "parent"
+            //                 }]
+            //             }
+            //         }
+            //     }
+            // }
+        });
+    });
+});
+
+describe('TiledeskChatbotUtil', function() {
+    describe('parseReply() of tdLinkSelf buttons', function() {
+        it('should return a link.SELF button', function() {
+            // const cbutil = new TiledeskChatbotUtil();
+            //const text = 'Intro text\n* Button with text  http://www.google.com';
+            const text = 'Intro text\n* Button with text link in self > http://www.link-in-self.com';
+            
+            console.log("parsing text:", text);
+            const reply = TiledeskChatbotUtil.parseReply(text);
+            console.log("reply:", JSON.stringify(reply));
+            assert(reply.message != null);
+            assert(reply.message.text != null);
+            assert.strictEqual(reply.message.text, 'Intro text');
+            assert.strictEqual(reply.message.type, TiledeskChatbotUtil.TYPE_TEXT);
+            assert(reply.message.attributes != null);
+            assert(reply.message.attributes.attachment != null);
+            assert(reply.message.attributes.attachment.buttons != null);
+            assert(reply.message.attributes.attachment.buttons.length == 1);
+            assert.strictEqual(reply.message.attributes.attachment.buttons[0].type, TiledeskChatbotUtil.TYPE_BUTTON_URL);
+            assert.strictEqual(reply.message.attributes.attachment.buttons[0].value, 'Button with text link in self');
+            assert.strictEqual(reply.message.attributes.attachment.buttons[0].link, 'http://www.link-in-self.com');
+            assert.strictEqual(reply.message.attributes.attachment.buttons[0].target, TiledeskChatbotUtil.TARGET_BUTTON_LINK_SELF);
+            // MESSAGE:
+            // {
+            //     "message": {
+            //         "text": "Intro text",
+            //         "type": "text",
+            //         "attributes": {
+            //             "attachment": {
+            //                 "type": "template",
+            //                 "buttons": [{
+            //                     "type": "url",
+            //                     "value": "Button with text link in self",
+            //                     "link": "http://www.link-in-self.com",
+            //                     "target": "self"
             //                 }]
             //             }
             //         }
@@ -385,6 +476,47 @@ describe('TiledeskChatbotUtil', function() {
             assert.strictEqual(reply.message.attributes.attachment.buttons[0].type, TiledeskChatbotUtil.TYPE_BUTTON_ACTION);
             assert.strictEqual(reply.message.attributes.attachment.buttons[0].value, 'Action Button with text');
             assert.strictEqual(reply.message.attributes.attachment.buttons[0].action, 'ACTION-CALLBACK-NAME');
+            assert.strictEqual(reply.message.attributes.attachment.buttons[0].show_reply, true);
+            // MESSAGE:
+            // {
+            //     "message": {
+            //         "text": "Intro text",
+            //         "type": "text",
+            //         "attributes": {
+            //             "attachment": {
+            //                 "type": "template",
+            //                 "buttons": [{
+            //                     "type": "action",
+            //                     "value": "Action Button with text",
+            //                     "action": "ACTION-CALLBACK-NAME",
+            //                     "show_reply": true
+            //                 }]
+            //             }
+            //         }
+            //     }
+            // }
+        });
+    });
+});
+
+describe('TiledeskChatbotUtil', function() {
+    describe('parseReply() of tdIntent buttons', function() {
+        it('should return an Intent button (aka Action button with show_reply = true)', function() {
+            const text = 'Intro text\n* Action Button with text tdIntent:INTENT-NAME';
+            console.log("parsing text:", text);
+            const reply = TiledeskChatbotUtil.parseReply(text);
+            console.log("reply:", JSON.stringify(reply));
+            assert(reply.message != null);
+            assert(reply.message.text != null);
+            assert.strictEqual(reply.message.text, 'Intro text');
+            assert.strictEqual(reply.message.type, TiledeskChatbotUtil.TYPE_TEXT);
+            assert(reply.message.attributes != null);
+            assert(reply.message.attributes.attachment != null);
+            assert(reply.message.attributes.attachment.buttons != null);
+            assert(reply.message.attributes.attachment.buttons.length == 1);
+            assert.strictEqual(reply.message.attributes.attachment.buttons[0].type, TiledeskChatbotUtil.TYPE_BUTTON_ACTION);
+            assert.strictEqual(reply.message.attributes.attachment.buttons[0].value, 'Action Button with text');
+            assert.strictEqual(reply.message.attributes.attachment.buttons[0].action, 'INTENT-NAME');
             assert.strictEqual(reply.message.attributes.attachment.buttons[0].show_reply, true);
             // MESSAGE:
             // {