+++ tiledesk-client/test/test.js
@@ -8,6 +8,8 @@ var jwt = require('jsonwebtoken');
 // FIRST PROJECT (OR ANOTHER PROJECT) DATA (ID AND NAME)
 const EMAIL = "test@tiledesk.com"; // first user
 const PASSWORD = "testtest";
+const EMAIL2 = "test2@tiledesk.com"; // secondo user
+const PASSWORD2 = "testtest";
 const PROJECT_ID = "6011eafd51245600345cdf72"; // first user project
 const PROJECT_NAME = 'First Test Project' // first project name
 const PROJECT_SECRET = "f42349dd-882c-40c8-92ad-7f385f073dce";
@@ -118,8 +120,8 @@ describe('TiledeskClient', function() {
               };
               console.log("externalUser", externalUser);
               var signOptions = {                                                            
-                subject:  'userexternal',                                                                 
-                audience:  'https://tiledesk.com/projects/' + PROJECT_ID                                             
+                subject:  'userexternal',
+                audience:  'https://tiledesk.com/projects/' + PROJECT_ID
               };
               var jwtCustomToken = "JWT " + jwt.sign(externalUser, PROJECT_SECRET, signOptions);
               
@@ -583,6 +585,60 @@ describe('TiledeskClient', function() {
                     assert(result.text === text_value);
                     done();
                 });
+            }
+            else {
+                assert.ok(false);
+            }
+        });
+    });
+});
+
+describe('TiledeskClient', function() {
+    describe('sendDirectMessage() by anonymous', function() {
+        it('sends a direct message ANONYM_USER_TOKEN to on-the-fly-anonynous user, "project_id" & "token" are "options" parameters', function(done) {
+            const tdclient = new TiledeskClient({
+                APIKEY: APIKEY,
+                APIURL: API_ENDPOINT,
+                log: true
+            });
+            if (tdclient) {
+              assert(tdclient != null);
+              tdclient.authEmailPassword(
+                EMAIL2,
+                PASSWORD2, 
+                function(err, resbody) {
+                    if (!err && resbody) {
+                        console.log("resbody:", resbody)
+                        assert(resbody.token != null);
+                        assert(resbody.user != null);
+                        assert(resbody.user._id != null);
+                        const text_value = "test message";
+                        const msgJSON = {
+                            "senderFullname": "Guest 1",
+                            "recipient": resbody.user._id,
+                            "text": text_value
+                        }
+                        // console.log("Sending message to REQUEST-ID:", request_id);
+                        tdclient.sendDirectMessage(msgJSON, function(err, result) {
+                            console.log("Message sent:", result)
+                            if (err) {
+                                console.error("An error occurred:", err);
+                                assert(err === null);
+                            }
+                            assert(result != null);
+                            assert(result.text === text_value);
+                            done();
+                          },
+                          {
+                              project_id: PROJECT_ID,
+                              token: USER_TOKEN
+                          });
+                    }
+                    else {
+                        assert.ok(false);
+                    }
+                }
+              );
             }
             else {
                 assert.ok(false);